pyproject.toml
[tool.poetry]
name = "garmin-to-fittrackee"
version = "0.2.1"
description = "A simple script to synchronize garmin to fittrackee. Inspired by https://github.com/jat255/strava-to-fittrackee"
authors = ["Dryusdan <contact@dryusdan.fr>"]
license = "GPL-3"
readme = "README.md"
packages = [{include = "garmin_to_fittrackee"}]

[tool.poetry.scripts]
garmin2fittrackee = "garmin_to_fittrackee.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
garminconnect = "^0.2.12"
pendulum = "^3.0.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
isort = "^5.13.2"



[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"
pytest-cov = "^4.1.0"
coverage = {extras = ["toml"], version = "^7.4.0"}

[[tool.poetry.source]]
name = "gitea"
url = "https://git.dryusdan.fr/api/packages/Dryusdan/pypi"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
docstring-code-format = true

[tool.coverage.run]
omit = [
    "tests/*",
]
[tool.coverage.report]
fail_under = 90
